refactor especially player handling to improve semantic understanding of whats happening
this means to "localize" (cant come up with a better word) functionality
to make sure all (well, most) of player interactivity and handling is
actually handled in the player-related stuff.
The "reference_triangle" is necessary to make the rotation work
properly, but doesnt have to be a whole ass object for itself, so its
been refactored to be part of the ship members.